Update patch set 1
Patch Set 1:

let me try and explain....

two dB is nothing. I'm an audiophile. a background first...

89 db is no gain. it is recorded as it comes.

each 9 db is a doubling of volume to the human ear

First off there are probably different laws in different parts of the world as to how loud something should be. Especially in Europe. That doesn't mean its going to perform different. it is just that, volume differences.

Now here is what I think is going on. Samsung probably has either a lib or in source something in place that already does the gain boosting for recording video. Now all we are doing is doing it in the mixer path. 106 - 89 = 17. It's not a random number I picked. If you look other high gain recording mics use that value.

Please can we just friggin merge this. its stupid to debate about 2 db differences. the human ear doesn't notice the change

Patch-set: 1