Update patch set 2
Patch Set 2: Code-Review+1

been testing for awhile now. No issues. This shouldn't be confused with my earlier patch to the kernel related to the ES705 chip. That patch was something different. It was for soley MIC bugs. This I had in place for a long time in my unofficial builds it must have just slipped through during the merge...

Patch-set: 2
Label: Code-Review=+1