Update patch set 2
Patch Set 2: Code-Review+1

Hi, please see
http://review.cyanogenmod.org/#/c/75577/
http://review.cyanogenmod.org/#/c/75578/

I added KeyDiasabler.java on samsung/hardware project as I guess is its natural place to be.
Then I added to device/samsung/klte-common only the change regarding the chown+chmod of the interface file.

Moreover, the tests done by KeyDisabler.java against the hardware interface file are more exhaustive (in comparison to this one) so if the system can't write/read on it, will not even show the option to the user in Settings. For instance, if other devices has this interface exposed in other filename or simply with wrong permissions/ownership.

Side Effects:

Physical Home button will still be operative, so its a redundancy to the navbar button. Nothing else. Not bad or weird behavior here.

I prefer navbar so all my devices has the same interface, which is at the end the reason AOSP encourages manufacturers to stop sing hardware buttons in opposition to the navbar. I guess other users could love it too.

Also, Cyanogenmod offer this option in other devices such as the S4 (jflte). Not offering this option to cm users of klte* leads to an inconsistency between similar devices (IMHO).

I'm using navbar in kltexx since 4 days ago and I'll still doing it regardless is in official branch or not.

Patch-set: 2
Label: Code-Review=+1