Update patch set 1
Patch Set 1:

according to the comments in the init script... (off the top of my head) is something to do with it not waiting long enough to try and mount those partitions. and if it fails once all further late class services will not start which makes you stuck on the boot animation with no logcat or anything. I'll try and find the comments.

FYI Sharrief used this exact patch on HLTE and it is merged and confirmed working

Patch-set: 1