Update patch set 1
Patch Set 1:

its been tested on more then one device that was giving this issue and fixed both and can be brought back by simply removing this change. An example is klteusc. on its NCB bootloader the device wont boot without this change. if you update it to NG3 it will. BUT not all devices have this option due to carriers being lazy. Like Verizon the Dev edition only has the bootloader it comes with. NO update. and if you flash the retail update then your bootloader is locked. Thus this is the ONLY way to get those devices to boot

Patch-set: 1