Update patch set 2
Patch Set 2:

ok i'll give it a compile now. Thing is i personally don't have the issue but i have two friends/testers who do. one is ktoonsez and another just a friend so it might take 12 hours or so before i know for sure if it works.. but i'll know in about 35 mins if it breaks anything but it shouldnt...

Patch-set: 2