Update patch set 1
Patch Set 1:

not sure about this. We were having Bluetooth Share FCs when devices would leave the range of BT and come back. I don't know if it was these or upstream changes that fixed it because there was a lot going on with bT at the time of this commit

Patch-set: 1
Reviewer: Matt Filetto <2814@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0