Update patch set 16
Patch Set 16:

To clarify: on CM nightlies I receive the calls, phone rings(with an old fashion ring tone) but, never prompted to accept or decline the call and it continues to ring for ever even after the caller has reached Voice-mail and has hung up. Only way to stop the ringing is by rebooting.

Patch-set: 16