Update patch set 18
Patch Set 18:

I can agree with that. I mean I've been running an old unofficial that has this fix... So far I haven't noticed any major issues with the exception of Bluetooth crashing when it's turned off. I'm not sure if that's related to this or not. I'd love to be able test, but can't seem to get repo sync to stop failing. I don't want to cause drama here, I just want a semi working device.. Being able to make calls and use data is way more important to me then Bluetooth crashing.. *shrug*

Patch-set: 18