Update patch set 1
Patch Set 1:

I want to see the behavior without this patch. Just grab a log when playing the volume up/down sound, or when a notification comes in. Those should be low latency use cases, I want to see what mixer and use cases are being selected for those.

Patch-set: 1