Update patch set 1
Patch Set 1: Code-Review+1

Working just fine so far on my device.  I hope giving credit in the commit message is ok.  If not, let me know if I need to change the author.  The original bacon commit was too different from what we had already for this device to do a clean cherry-pick.

Patch-set: 1