Update patch set 1
Patch Set 1:

Sorry to say but I have no idea how to download the code to my computer or insert these lines into it and try it on my klte. I can imagine is that it will most likely work, just like for that other device. If I knew how, I would've implemented this a while ago and submitted it for consideration. Maybe an experienced contributor could take a moment and copy and paste these 4 lines and commit the change. If it doesn't work it can always be reverted.

Patch-set: 1