Update patch set 2
Abandoned

After using this for a couple days, it's kind of wonky.  Ethan has a point, without pickup sensor, it's just kind of frustrating.

It works, but not full enough to really hope to enable.

I only started looking at it because the QS tile showed up recently.

Patch-set: 2
Status: abandoned