Update patch set 4
Patch Set 4:

I've got a working test using the proximity sensor.  I directly modified the doze service in systemUI and added some necessary parameters to statusbar/phone/DozeParameters.  It's working for me right now, but I don't know if it'll be an acceptable solution.  I did it more as proof of the concept for using the proximity sensor.

Patch-set: 4