klte-common: fix performance profiles
Fix kgsl-3d0 devfreq governor location

Enable the cpu before setting the cpufreq governor or the cpu?/cpufreq
directory won't exist and the governor won't change (the governors would
stay as "performance" after switching to the "performance" profile and
later going back to "balanced"). The hotplug daemon must be off while
enabling the CPU or it could disable them again before we get a chance
to change the governor.

For the powersave profile to stick to interactive governor as
conservative keeps the frequency around 1ghz with screen on (it only
goes down to 300mhz with screen off), but allow only 1 CPU with the
powersave profile (that is still usable with interactive). That gives me
more confidence that can save power than to keep one CPU enabled at 1ghz
at all times and when load increases multiple of them and none at
300mhz. Disable the hotplug daemon for the powersave case like for the
performance case.

Change-Id: I4e9c0ddc1d28617b9200a92adbf410d87302acbb