Update patch set 1
Patch Set 1:

There's nothing to change for this in the kernel.  The proximity sensor is already a wake up sensor, so this basically just adds a listener for it's broadcast.

If new wake up sensors are ever added or discovered, the service should be pretty easy to modify buy just adding a new sensor class that implements the IKlteSensor interface, and adding it to the service startup

Patch-set: 1