Update patch set 1
Patch Set 1:

Ethan, For the overlay, are you referring to the systemUI one?  That's already defaulted to false, so I didn't think it was necessary.

Regarding the package definition.  To get anything in settings, the only device one's I see are for Bluetooth and Touchscreen gestures.  This doesn't really fit into those categories.  Is there another one that I'm missing?

Patch-set: 1