Update patch set 1
Patch Set 1:

Good point about the pickup sensor overlay. As for the proximity sensor class, the reason most of the other implementations had a separate sensor class is because we needed to encapsulate additional behavior in the sensor classes. The ProximitySensor class doesn't provide anything the existing Sensor class doesn't. If you had it as an inner class you'd be able to avoid having a static context as well.

Patch-set: 1