Update patch set 1
Patch Set 1:

Yeah, I already re-coded the context part. Basically the prox sensor class just takes it into the constructor and holds it as a private variable to pass back to the service to do the doze pulse.

If there is a place to create a settings interface to configure the prox sensor on off maybe, those checks could go into the prox sensor class to make it's separation more desirable.

Did you see my other question regarding exposing some kind of settings screen?  I wasn't sure that there's an easy way to do it with existing functionality.

Patch-set: 1