Update patch set 1
Patch Set 1:

Inserting a newly formatted usb-stick causes the Android/data directory to be created and populated with a couple of running apps' names (but no data, of course).

This happens both with *and without* the noemulatedsd in my fstab.  I would guess because the storage_list.xml lists it as not emulated.

Patch-set: 1