Update patch set 5
Patch Set 5:

So, do we not want an option to allow for the out of pocket doze?

You removed the options, but left the checks.  That's fine if we just assume no one wants to have that choice.  But there's some leftover stuff regarding the pocket code that I can clean up if that's the case.

Other than that, I get the basic refactor.  It was to crush my dreams of ever having more sensors enabled :(

Patch-set: 5