Update patch set 3
Patch Set 3:

cat <existing file> <non-existing file> works in the sense that the existing file is still written, but it exits with a non-zero exit code. This is in particular a problem if the /last/ cat involves a non-existing file, because make fails in that case.

Patch-set: 3
Reviewer: Phillip Berndt <12931@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0