Update patch set 1
Patch Set 1:

The only reason to set it to powersave before shutting them off is to be sure they're off with mhz at the lowest freq. It should not matter, it's paranoid... because I don't know exactly which part of the cores are powered off, it felt safer that way.

About using interactive instead of conservative in lowpower profile, conservative just doesn't keep the CPU at the lowest freq when the phone is idle, so it doesn't look good. Reducing the cores that can get online in lowpower profile, should be more effective so we won't wake cores for a little while (false positives).

Overall the patch looks important as when switching to lowpower or to balanced, the governor could stay in "performance" mode leading to major battery leak (especially while the screen is on).

I recall while running waze just touching the screen of my phone felt like burning my finger, now it stays cool, I don't know if this is the change that made the difference as I haven't checked manually the governor while I was running waze that day it was overheating, but I don't exclude it.

I recommend it for cm12.0 too.

Patch-set: 1