Update patch set 2
Patch Set 2: Code-Review-1

This is not good, you can't leave cpus enabled (no matter if forced on, or with hotplug) with the powersave governor. With this change the phone can run much slower (almost unusable) than with just 1 cpu on. The priority doesn't go to cpu0, if a task is running in cpu1, it will stay there to avoid MM switches and it'll keep computing there (eventually it may change but it may move to cpu2 not to cpu0).

All cpus are equal as far as the scheduler is concerned.

There's a bias perhaps on the cpu0 still as if they're all idle and a new task has been started, it may pick that, but previously running tasks won't have bias.

The removal of echo 0 > cpu0/enabled should be ok because if all others go down, we'd be guaranteed we're running on cpu0 and it's on.

Patch-set: 2