Update patch set 7
Patch Set 7:

> Patch Set 7: Code-Review+1

Everything are blow if(isLollipopRadio), so it don't need to tset for kk blob and I have tested on kltechn.
It spend many days to review it and I can't wait to merge it. Many people are wait for it.
I think I should merge common code to kltechnduo device so that I can change it by myself and will not waste my time.

Patch-set: 7