Update patch set 2
Patch Set 2:

Honestly it obviously doesn't affect performance or anything like that.  And I get the skepticism in adding this, I can't argue with that.

It does however make a transition from stock firmware to cm easier for the user, so they don't immediately think their signal strength decreased.

And just for my 2 cents, it's just for this device, absolutely no affect on other devices, so what's the harm (besides opening it up for other's to try and copy)

Patch-set: 2