Update patch set 2
Patch Set 2:

This for sure gives the right options M has for the sdcard, but if you try to format as internal storage, you can't boot with the sdcard in any longer. Guess it's that whole klte won't work with encrypted sdcard thing.  Definitely mounts properly now though, just a warning on formatting as internal

Patch-set: 2