Update patch set 1
Patch Set 1:

I see what's going on here. I think your dtb generating rule in your kernel is the zImage-dtb target. Generating a regular zImage is probably not creating dtbs for you.

My problem is that generating zImage-dtb gives you the concatenated kernel+dtb, and feeding that to mkbootimg in addition to the --dt dt.img parameter needlessly increases boot.img size. For klte, this results in a boot.img that's too large. klte's kernel is generating .dtb files without the zImage-dtb target, so that part isn't an issue for me.

Patch-set: 1