Update patch set 5
Patch Set 5:

I'm giving it a last try. i make a new build with the gps-patch for vendor, device and native for sensor and two uicc-patches. i will wipe the whole phone and patch without installing supersu.

which recovery are you using patrcik?

maybe we should create a bloochat or so....

Patch-set: 5