Update patch set 1
Patch Set 1:

I'm not sure about it either.  But the problem started right after applying the reverted patch, and it's now gone.  No one noticed it on dirty flashing since these permissions are set on files inside /data.  But once you format /data, and flash clean (or factory reset), the problems arise.

I thought it was weird too, but I can guarantee this fixes it.  Two builds side by side with only this different, and it's all good with this revert

Patch-set: 1