Update patch set 1
Patch Set 1:

Ethan, On a clean flash with the original patch still implemented, can you rotate to landscape on a sprint phone, or see the incall screen after making a call?

We noticed this happening on g900t and g900p.  g900f users didn't seem to be affected.

Patch-set: 1