Update patch set 2
Patch Set 2:

if this belongs to all qc feel free to move it. but this policy is really present in many devices so nobody else had ever had this thought maybe because if charger is not linked with healthd like it is not yet in some devices in init.qcom.rc it would not have any effect to them.

Patch-set: 2