Update patch set 3
Patch Set 3:

Black bars are just not transparent and as you (and people on xda) said, the final picture comes out correctly so why no merge? It is in the nature of nightlies that not everything works correctly out of the box. Also, patch to fix black bars in 16:9 mode has abandoned status as of now so waiting for it could take a while.

Patch-set: 3