Update patch set 1
Patch Set 1: Code-Review+1

tested working my self.

just a lil heads up, in some ocations, when and invalid scan is made you have to let the acreen goes black and then power it on to reset the finger entry.

example 

i swipe the finger wrong and a message says try again, i re-swipe and nothing happens, i ether have to bring up the option for pin code or turn the screen black and power on to re-swipe again.

Patch-set: 1