Update patch set 1
Patch Set 1:

That app setup seems fragile. Without any running service (let alone a foreground one), the system can kill that app at any given time.Also, what's the purpose of the screen receiver? The fingerprint HAL API expects the sensor to be able to wake up the device.

Additionally, seeing that the sensor seems to be implemented in the TEE (right?), did you consider/check reverse engineering the TEE protocol instead of using that weird vendor setup? For my device the TEE protocol was pretty simple and understandable, so it was easier to just make the HAL talk to the TEE directly.

Patch-set: 1