Update patch set 3
Patch Set 3:

Alan J. Ortiz the other commits on Mokee for klte kernel show for Chinese klte, but the first one dated Feb 17th says update fingerprint drivers which applies to all devices that use klte as a base. One thing that I forgot to mention, the fingerprint still works without the kernel merge, was just mentioning it in case people need it as a reference to fix any bugs.

Patch-set: 3