Update patch set 5
Patch Set 5:

No, I don't. 100kB buffers sound very (overly?) large, but I don't see a problem per we in this size for deciphering the protocol. If the protocol really uses that buffer size for whatever reason (e.g. interpreting the bitmaps in user space or whatever), then it's really going to be hard to decipher it though.

Patch-set: 5