Update patch set 6
Patch Set 6: Code-Review+1

this is working exelent for now, ill say marge it.

this can be cleaned up latter in newer builds.

this is not half baked this is hard work... no one here taked the task to acomplish this feature no one and hey... its now here thanks to this man sended by God to make this feature aviable on various cm13.0 roms.

Patch-set: 6