Update patch set 3
Patch Set 3:

what is this suppose to fix. if its your "Wifi issue" its not going to do anything... I'm telling you man. its a app you have. Pat and myself and plenty of others have never been able to reproduce. and sensorhub was removed for a reason. its not present in LL libs

Patch-set: 3