Update patch set 3
Patch Set 3: Code-Review+1

I did the clean build and flashed it with Deltadroid gapps like my last build and after 10 minutes with the same WiFi settings as before with same network and no drops. Not exactly sure how this fixed it and Matt you did tell me to do a complete factory reset without gapps but wanted to try this first as the only change and it worked.

Patch-set: 3
Label: Code-Review=+1