Update patch set 7
Patch Set 7:

Never mind my earlier post, I just tried using the sensor with stock and it still does not work. The hardware is apparently broken.

I'd suggest that you display an error message if the device reports an out-of-order status (Which it does; I see init: ret=360 in logcat.), instead of silently failing.

Patch-set: 7