Update patch set 3
Patch Set 3: Code-Review-1

after some more testing i also think that the errors haven't changed. the most problem is that gcm fails to whitelist the apps after gcm noticed that a push message arrived. but i also think that we have some selinux problems which causes wifi problems. when i set selinux to permisssive the wifi works much better. and there are some avc in cat that should be fixed so at this moment i think we don't need this.

Patch-set: 3