Update patch set 3
Patch Set 3: -Code-Review -Verified

I don't think this should be merged for that reason but for another. this hal was present in our cm-12.1 builds and we had no issues. When I updated the libs i completely missed it, I thought it was deleted. but if its there it should be added. But its not related to their wifi bug (which i cannot reproduce(. BUT i do know the general description of doze is "it uses all the sensors to see if the phone is in use" so MAYBE its a side effect but I dont see any evidence to support that

Patch-set: 3