Update patch set 3
Patch Set 3:

It was deleted by Ethan Ehen he introduced LP Sensors but i'm with you AS it existiert in stock and we used it in LP we should merge and keep it in our builds. But it would really be interesting decompile the so and see what it does.

Patch-set: 3