Update patch set 3
Patch Set 3:

I don't know what this patch is responsible for but I was trying unofficial Resurrection Remix which is based on CM13 with this patch integrated and Wifi stays connected when device is in DeepSleep most of the time. Wifi signal lost is still present but much less than before.

Patch-set: 3