Update patch set 2
Patch Set 2:

As a remark, I've found a transition that still produces a crash: Fast, continuous toggling of HDR mode in the camera app without taking photos eventually crashes the camera for me.

Not a serious issue, because that's not a common use case, but it's something to keep in mind for later.

Patch-set: 2
Reviewer: Phillip Berndt <12931@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0