Update patch set 1
Patch Set 1:

I originally started with the new kernel and needed this to get it to start to build. It ended up bombing in selinux around the required patches, so ended up saying "screw it" and went back to the old kernel with the compat patches.

Looks like we'll eventually need it, at least it's ready to repopick...

Patch-set: 1