Update patch set 1
Patch Set 1:

if the change to 10 fixes the flickering we will keep it until someone fixes the failure in hardware. i can't reproduce the bug so i have to look at forums and i didn't notice any complains about flickering since the original patch was merged.

it is the same way we did with hdr. disabling features until hardware works properly or any better way without bugs are found.

Patch-set: 1