Update patch set 1
Patch Set 1:

The only way to fix this isssue is to replace the hardware.
The problem is confirmed to also occur on stock Touchwiz, so this workaround should not be hardcoded in the software as it penalizes most of the users.

The best and cleanest way to deal with this would be to let the user adjust the minimum brightness level he wants, as it was done in good old time of CM. This would allow to keep the standard default value of 2 while letting users facing flickering issue deal with higher value to avoid their problem (something they can already do themselves by using some Xposed modules).

Patch-set: 1