Update patch set 1
Patch Set 1:

I share the opinion 10 is not low enough but we have to as much stabilty as possible so we have to keep it linke this to fix flickering for affected users. Also most other msm8974 devices uses these values. And as Paul said cm must work without set something via exposed. Plesse be patient guys and understand why we keep it linke this. Sry

Patch-set: 1