Update patch set 1
Patch Set 1:

I would be willing to consider writing some code to allow users to change this value themselves without using Xposed if that would help, there's a few things I have been meaning to patch and this is as good a place as any to start. I get that you have to support people with issues and have a lot on your plate. Thanks for all the great work you have done so far, I understand where you're coming from.

Patch-set: 1