Update patch set 1
Patch Set 1:

I think the logic here looks good, but now after cherry-picking the patch I wonder if we need to correct our vendor files a bit.
macloader identifies my device as murata. There is no _murata sets, but there is a muratafem1. macloader wouldn't know what to do with a muratafem1, kernel source says muratafem1 is different from everything else and that murata is the default "a0". The _a0 files are the same as the un-appended files and they say they are for "SEMCO SWB-B80 Rev031 module". WTF, over?

Patch-set: 1