Update patch set 1
Patch Set 1:

(1 comment)

Looks reasonable, but I'm kind of curious what the kernel is looking for in /efs. Do you happen to have the denial captured?

Not required, but it may be a good idea to start adding the denial we are fixing as a comment to the change record so that we can remind ourselves what we were fixing.

Patch-set: 1