Update patch set 9
Patch Set 9:

> for the reasons you have it doesn't interest me what you think
 > about. if you don't have any others reasons than you have been
 > crucified for your change that changes behaviour an other way than
 > my one and also never tested how dark you can turn your phone
 > without and without this patch you also don't ned to -1 cr if there
 > is not any bug in new code.

The minimum brightness and the minimum brightness auto brightness can be set should be the same, it's bad UX otherwise. Also, I can be opposed to a change even if it's not a bug, but I can't stop you from submitting it, since I don't have -2, nor would I stop you from making a decision on a device I don't own. But I would like to note my opposition to this. My main problem is that the UX would be awful with autobrightness setting one minimum brightness, but being able to lower the brightness below that, because that is inconsistent and jarring

Patch-set: 9