Update patch set 4
Patch Set 4:

(4 comments)

i also thought this way but there are more /dev/ in init.qcom maybe because it is handled like this in stock. so we have to decide if we Keep the dev in here like stock or we move all them to uevent. I would prefer moving them for a better transparency.

Patch-set: 4
Label: Verified=0