Update patch set 4
Patch Set 4:

Timezone just kicked back to GMT, even with this implementation. :/

I wish it were more deterministic so that it were easier to capture relevant logs. It's easy enough to dump a logcat, but finding the event somewhere in up to 18H of logs is the proverbial needle in a haystack.

Patch-set: 4
Label: Verified=0