Update patch set 11
Patch Set 11:

The main problem with PS11 is that it's not fixing the Parcel, so all the changes done by this subclass are lost once super.processUnsolicited() is called. If replace "break" with "return" in the switch, PS11 might work, but fixing the Parcel is simpler and better IMO.

p.setDataSize(p.dataPosition()) shouldn't be necessary in my patch, but I added it anyway (and hopefully is resizing the parcel correctly).

Patch-set: 11
Reviewer: Gabriele M <13897@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0