Update patch set 8
Patch Set 8:

Did you pull the blobs from stock? And are you testing all the effects in AudioFX?
Also, this is behaving sanely for normal playback, but I need to test all audio usecases (calls, recording, etc), so this is going to take a while to get in. Right now I'm working on the low microphone issue, for which I'm on stock.

Patch-set: 8
Label: Verified=0