Update patch set 1
Patch Set 1:

In case it helps you to decide, with my patch it also works perfectly in OpenCamera.  OpenCamera doesn't use video-hdr, instead it sets scene-mode=hdr/auto for both photo and video, and the Samsung blob turns on rt-hdr by itself. It's just Snap that doesn't support scene-mode for video and instead wants video-hdr. I could have translated video-hdr to scene-mode in the wrapper, but since it's a user visible setting it could cause confusion, so instead I translated it to rt-hdr.

Patch-set: 1